+++ commands/root.go
@@ -182,7 +182,7 @@ __custom_func() {
 __awless_region_list()
 {
     cur="${COMP_WORDS[COMP_CWORD]#*=}"
-    regions="us-east-1 us-east-2 us-west-1 us-west-2 ca-central-1 eu-west-1 eu-central-1 eu-west-2 ap-northeast-1 ap-northeast-2 ap-southeast-1 ap-southeast-2 ap-south-1 sa-east-1"
+    regions="us-east-1 us-east-2 us-west-1 us-west-2 ca-central-1 eu-west-1 eu-central-1 eu-west-2 eu-west-3 ap-northeast-1 ap-northeast-2 ap-southeast-1 ap-southeast-2 ap-south-1 sa-east-1"
     COMPREPLY=( $(compgen -W "${regions}" -- ${cur}) )
 }
 
@@ -196,7 +196,7 @@ __awless_profile_list()
 __awless_profile_region_list()
 {
     cur="${COMP_WORDS[COMP_CWORD]#*=}"
-		regions="us-east-1 us-east-2 us-west-1 us-west-2 ca-central-1 eu-west-1 eu-central-1 eu-west-2 ap-northeast-1 ap-northeast-2 ap-southeast-1 ap-southeast-2 ap-south-1 sa-east-1"
+		regions="us-east-1 us-east-2 us-west-1 us-west-2 ca-central-1 eu-west-1 eu-central-1 eu-west-2 eu-west-3 ap-northeast-1 ap-northeast-2 ap-southeast-1 ap-southeast-2 ap-south-1 sa-east-1"
     profiles="$((egrep '^\[ *[a-zA-Z0-9_-]+ *\]$' ~/.aws/credentials 2>/dev/null; grep '\[profile' ~/.aws/config 2>/dev/null | sed 's|\[profile ||g') | tr -d '[]' | sort | uniq)"
     COMPREPLY=( $(compgen -W "${profiles} ${regions}" -- ${cur}) )
 }